+++ gcs/CHANGES.md
@@ -1,5 +1,7 @@
 ### 2.1.7 - 2020-XX-XX
 
+1.  Enable TCP keep alive by default.
+
 ### 2.1.6 - 2020-11-06
 
 1.  Increment Hadoop `FileSystem.Statistics` counters for read and write